+++ webapp/src/components/App/Singlestudy/explore/Modelization/BindingConstraints/BindingConstView/ConstraintFields.tsx
@@ -27,10 +27,10 @@ import SwitchFE from "../../../../../../common/fieldEditors/SwitchFE";
 import { useFormContextPlus } from "../../../../../../common/Form";
 import { useMemo, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { validateString } from "../../../../../../../utils/validationUtils";
 import Matrix from "./Matrix";
 import { Box, Button } from "@mui/material";
 import { Dataset } from "@mui/icons-material";
+import { validateString } from "@/utils/validation/string";
 
 interface Props {
   study: StudyMetadata;
