+++ webapp/src/components/App/Singlestudy/explore/TableModeList/dialogs/TableTemplateFormDialog.tsx
@@ -23,11 +23,9 @@ import SelectFE from "../../../../../common/fieldEditors/SelectFE";
 import StringFE from "../../../../../common/fieldEditors/StringFE";
 import { getTableColumnsForType, type TableTemplate } from "../utils";
 import { TABLE_MODE_TYPES } from "../../../../../../services/api/studies/tableMode/constants";
-import {
-  validateArray,
-  validateString,
-} from "../../../../../../utils/validationUtils";
 import { useMemo } from "react";
+import { validateArray } from "@/utils/validation/array";
+import { validateString } from "@/utils/validation/string";
 
 export interface TableTemplateFormDialogProps
   extends Pick<
