+++ webapp/src/components/App/Singlestudy/explore/Xpansion/Candidates/CreateCandidateDialog.tsx
@@ -25,7 +25,7 @@ import Fieldset from "../../../../../common/Fieldset";
 import SelectFE from "../../../../../common/fieldEditors/SelectFE";
 import NumberFE from "../../../../../common/fieldEditors/NumberFE";
 import { SubmitHandlerPlus } from "../../../../../common/Form/types";
-import { validateString } from "../../../../../../utils/validationUtils";
+import { validateString } from "@/utils/validation/string";
 
 interface PropType {
   open: boolean;
